ChatPDF-Server/requirements.txt
fastapi
groq
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-text-splitters
PyMuPDF
python-dotenv
uuid
